requirements.txt
feedparser
git
ics==0.5
libopenjp2-7-dev
libtiff5
pathlib
Pillow==5.3.0
pip3
pyowm
python3-pip
python-rpi.gpio-dbgsym
python3-rpi.gpio
python-rpi.gpio
python3-rpi.gpio-dbgsym
python3-spidev
python3-numpy
pytz
tzlocal
pyyaml
